Double to float
Melhorar a performance, o programa está muito lento devido a quantidade
de memória utilizada.

+++ HMax_Class/s1th.h
@@ -19,8 +19,8 @@ class S1Th : public QThread
     Q_OBJECT
 public:
     explicit S1Th(QObject *parent = 0);
-    explicit S1Th(cv::Mat image, std::vector<int>* sizes, std::vector<double>* lambda, std::vector<double>* sigma,
-                  std::vector<double>* gama, std::vector<double>* orientation, std::vector<cv::Mat>* filters, QObject *parent = 0);
+    explicit S1Th(cv::Mat image, std::vector<int>* sizes, std::vector<float>* lambda, std::vector<float>* sigma,
+                  std::vector<float>* gama, std::vector<float>* orientation, std::vector<cv::Mat>* filters, QObject *parent = 0);
 
     std::vector<S1_T>* gaborFilterResult;
     std::vector<cv::Mat>* filters;
@@ -30,10 +30,10 @@ public:
 private:
     cv::Mat image;
     std::vector<int>* sizes;
-    std::vector<double>* lambda;
-    std::vector<double>* sigma;
-    std::vector<double>* gama;
-    std::vector<double>* orientation;
+    std::vector<float>* lambda;
+    std::vector<float>* sigma;
+    std::vector<float>* gama;
+    std::vector<float>* orientation;
     void run();
 };
